move default impl to xtext package
